rag-backend/requirements.txt
fastapi
uvicorn
pdfplumber
pymupdf
langchain-core 
langchain-community
langchain-core
torch
transformers
tokenizers
numpy
faiss-cpu
python-multipart
sentence-transformers
python-docx
